buyer_orchestrator_agent/requirements.txt
httpx>=0.24.0
asyncio
python-dotenv
uvicorn
click
logging
# A2A dependencies - these should be available from your ADK installation
# a2a
# google.adk
